+++ uCOS-II_Wrapper/examples/task_example.c
@@ -7,9 +7,11 @@ static OS_STK task_stack [TASK_SIZE];
 
 static void task(void *p_arg)
 {
+    OS_STK_DATA stk;
     while(1)
     {
-        rt_kprintf("TASK\r\n");
+        OSTaskStkChk(OS_PRIO_SELF, &stk);
+        rt_kprintf("task stack free: %d\r\n", stk.OSFree);
         rt_thread_delay(500);
     }
 }
@@ -17,8 +19,8 @@ static void task(void *p_arg)
 void task_example (void)
 {
 #if OS_STK_GROWTH == 0u      
-    OSTaskCreateExt(task,0,&task_stack[0],TASK_PRIO,0,&task_stack[TASK_SIZE-1],TASK_SIZE,0,0);
+    OSTaskCreateExt(task,0,&task_stack[0],TASK_PRIO,0,&task_stack[TASK_SIZE-1],TASK_SIZE,0,OS_TASK_OPT_STK_CHK|OS_TASK_OPT_STK_CLR);
 #else
-    OSTaskCreateExt(task,0,&task_stack[TASK_SIZE-1],TASK_PRIO,0,&task_stack[0],TASK_SIZE,0,0);
+    OSTaskCreateExt(task,0,&task_stack[TASK_SIZE-1],TASK_PRIO,0,&task_stack[0],TASK_SIZE,0,OS_TASK_OPT_STK_CHK|OS_TASK_OPT_STK_CLR);
 #endif
 }