+++ ChangeLog.md
@@ -36,6 +36,11 @@
 
 - 实现`OSTaskChangePrio` `OSTaskResume` `OSTaskSuspend` `OSTaskStkChk`函数, task.c函数全部完成
 
+### 2020-10-05
+
+- 修复`OSTaskCreateExt`函数堆栈增长方向兼容RT-Thread的问题
+- 修复`OSTaskStkChk`函数计算任务堆栈大小错误的问题
+
 
 
 # 版本